Reorganize packages for Dialogs
